adding exceptions

+++ src/main/java/com/example/graphqltraining/resolver/Mutation.java
@@ -3,6 +3,9 @@ package com.example.graphqltraining.resolver;
 import com.example.graphqltraining.bean.Client;
 import com.example.graphqltraining.bean.Compte;
 import com.example.graphqltraining.bean.Operation;
+import com.example.graphqltraining.exception.ClientAlreadyExist;
+import com.example.graphqltraining.exception.NotFoundException;
+import com.example.graphqltraining.exception.ValidationException;
 import com.example.graphqltraining.repository.ClientRepository;
 import com.example.graphqltraining.repository.CompteRepository;
 import com.example.graphqltraining.repository.OperationRepository;
@@ -32,10 +35,16 @@ public class Mutation implements GraphQLMutationResolver {
                                    String email,
                                    String address,
                                    Date dateNaissance) {
-        Client client = new Client(CNE, firstName, lastName, phoneNumber, email, address, dateNaissance);
-        client.setCode(CodeUtil.generateCode(10));
-        clientRepository.save(client);
-        return client;
+        if(clientDataIsValid(CNE, firstName, lastName, phoneNumber, email)){
+            if(clientRepository.findByEmail(email) != null
+                    || clientRepository.findByCNE(CNE) != null
+                    || clientRepository.findByPhoneNumber(phoneNumber) != null)
+                throw new ClientAlreadyExist("Client is already exist");
+            Client client = new Client(CNE, firstName, lastName, phoneNumber, email, address, dateNaissance);
+            client.setCode(CodeUtil.generateCode(10));
+            return clientRepository.save(client);
+        }else
+            throw new ValidationException("Send all the values");
     }
 
     @Transactional
@@ -47,21 +56,26 @@ public class Mutation implements GraphQLMutationResolver {
                                 String email,
                                 String address,
                                 Date dateNaissance){
-        Client client = clientRepository.getReferenceById(id);
-        client.setCNE(CNE);
-        client.setFirstName(firstName);
-        client.setLastName(lastName);
-        client.setPhoneNumber(phoneNumber);
-        client.setEmail(email);
-        client.setAddress(address);
-        client.setDateNaissance(dateNaissance);
-        return clientRepository.save(client);
+        if(clientDataIsValid(CNE, firstName, lastName, phoneNumber, email)){
+            Client client = clientRepository.getReferenceById(id);
+            client.setCNE(CNE);
+            client.setFirstName(firstName);
+            client.setLastName(lastName);
+            client.setPhoneNumber(phoneNumber);
+            client.setEmail(email);
+            client.setAddress(address);
+            client.setDateNaissance(dateNaissance);
+            return clientRepository.save(client);
+        }else
+            throw new ValidationException("Send all the values");
     }
 
     @Transactional
     public boolean deleteClient(Long id){
-        clientRepository.deleteById(id);
-        return true;
+        if (clientRepository.findById(id).isPresent()){
+            clientRepository.deleteById(id);
+            return true;
+        }else throw new NotFoundException("Client not found");
     }
 
     @Transactional
@@ -69,10 +83,15 @@ public class Mutation implements GraphQLMutationResolver {
                                 Date dateCreation,
                                 String typeCompte,
                                 Long idClient){
-        Client client = clientRepository.getReferenceById(idClient);
-        Compte compte = new Compte(solde, dateCreation, typeCompte, client);
-        compte.setNumero(NumeroUtil.generateNumero());
-        return compteRepository.save(compte);
+        if(!typeCompte.trim().equals("")){
+            if(clientRepository.findById(idClient).isPresent()){
+                Client client = clientRepository.getReferenceById(idClient);
+                Compte compte = new Compte(solde, dateCreation, typeCompte, client);
+                compte.setNumero(NumeroUtil.generateNumero());
+                return compteRepository.save(compte);
+            }else throw new NotFoundException("Client not found");
+        }else throw new ValidationException("Send all the values");
+
     }
 
     @Transactional
@@ -81,19 +100,27 @@ public class Mutation implements GraphQLMutationResolver {
                                Date dateCreation,
                                String typeCompte,
                                Long idClient){
-        Client client = clientRepository.getReferenceById(idClient);
-        Compte compte = compteRepository.getReferenceById(id);
-        compte.setSolde(solde);
-        compte.setDateCreation(dateCreation);
-        compte.setTypeCompte(typeCompte);
-        compte.setClient(client);
-        return compteRepository.save(compte);
+        if(!typeCompte.trim().equals("")){
+            if(compteRepository.findById(id).isPresent()){
+                if(clientRepository.findById(idClient).isPresent()){
+                    Client client = clientRepository.getReferenceById(idClient);
+                    Compte compte = compteRepository.getReferenceById(id);
+                    compte.setSolde(solde);
+                    compte.setDateCreation(dateCreation);
+                    compte.setTypeCompte(typeCompte);
+                    compte.setClient(client);
+                    return compteRepository.save(compte);
+                }else throw new NotFoundException("Client not found");
+            }else throw new NotFoundException("Compte not found");
+        }else throw new ValidationException("Send all the values");
     }
 
     @Transactional
     public boolean deleteCompte(Long id){
-        compteRepository.deleteById(id);
-        return true;
+        if (compteRepository.findById(id).isPresent()){
+            compteRepository.deleteById(id);
+            return true;
+        }else throw new NotFoundException("Compte not found");
     }
 
     @Transactional
@@ -101,10 +128,15 @@ public class Mutation implements GraphQLMutationResolver {
                                      float montant,
                                      String typeOpration,
                                      Long idCompte){
-        Compte compte = compteRepository.getReferenceById(idCompte);
-        Operation operation = new Operation(date, montant, typeOpration, compte);
-        operation.setNumero(NumeroUtil.generateNumero());
-        return operationRepository.save(operation);
+        if(!typeOpration.trim().equals("")){
+            if(compteRepository.findById(idCompte).isPresent()){
+                Compte compte = compteRepository.getReferenceById(idCompte);
+                Operation operation = new Operation(date, montant, typeOpration, compte);
+                operation.setNumero(NumeroUtil.generateNumero());
+                return operationRepository.save(operation);
+            }else throw new NotFoundException("Compte not found");
+        }else throw new ValidationException("Send all the values");
+
     }
 
     @Transactional
@@ -113,19 +145,39 @@ public class Mutation implements GraphQLMutationResolver {
                                      float montant,
                                      String typeOpration,
                                      Long idCompte){
-        Operation operation = operationRepository.getReferenceById(id);
-        Compte compte = compteRepository.getReferenceById(idCompte);
-        operation.setDate(date);
-        operation.setMontant(montant);
-        operation.setTypeOperation(typeOpration);
-        operation.setCompte(compte);
-        return operationRepository.save(operation);
+        if(!typeOpration.trim().equals("")){
+            if(operationRepository.findById(id).isPresent()){
+                if(compteRepository.findById(idCompte).isPresent()){
+                    Operation operation = operationRepository.getReferenceById(id);
+                    Compte compte = compteRepository.getReferenceById(idCompte);
+                    operation.setDate(date);
+                    operation.setMontant(montant);
+                    operation.setTypeOperation(typeOpration);
+                    operation.setCompte(compte);
+                    return operationRepository.save(operation);
+                }else throw new NotFoundException("Compte not found");
+            }else throw new NotFoundException("Operation not found");
+
+        }else throw new ValidationException("Send all the values");
     }
 
     @Transactional
     public boolean deleteOperation(Long id){
-        operationRepository.deleteById(id);
-        return true;
+        if (operationRepository.findById(id).isPresent()) {
+            operationRepository.deleteById(id);
+            return true;
+        }else throw new NotFoundException("Operation not found");
     }
 
+    private boolean clientDataIsValid(String CNE,
+                                       String firstName,
+                                       String lastName,
+                                       String phoneNumber,
+                                       String email){
+        return !CNE.trim().equals("") &&
+                !firstName.trim().equals("") &&
+                !lastName.trim().equals("") &&
+                !phoneNumber.trim().equals("") &&
+                !email.trim().equals("");
+    }
 }
